set set_default_dtype to double

+++ spateo/tools/interpolation_nn.py
@@ -207,7 +207,7 @@ class MainFlow(torch.nn.Module):
 
     def forward(self, t, x, freeze=None):
 
-        x_low = self.A(x.double()) if self.A is not None else x.double()
+        x_low = self.A(x) if self.A is not None else x
         e_low = self.h.forward(x_low)
         e_hat = self.B(e_low) if self.B is not None else e_low
